# feat: pass in formatters to IntlMessageFormat for perf