--- a/test/perf/index.tsx
+++ b/test/perf/index.tsx
@@ -8,6 +8,7 @@ import {
   FormattedMessage,
   // FormattedRelativeTime,
 } from '../../dist';
+import parser, {MessageFormatPattern} from 'intl-messageformat-parser';
 
 const suite = new Suite('renderToString', {
   onCycle: function(e: any) {
@@ -16,7 +17,7 @@ const suite = new Suite('renderToString', {
 });
 
 suite.on('error', function(e: any) {
-  console.log(e)
+  console.log(e);
   throw e.target.error;
 });
 
@@ -25,14 +26,12 @@ suite.add('<div>', function() {
 });
 
 suite.add('100 x <div/>', function() {
-  const divs = []
+  const divs = [];
   for (let i = 0, len = 100; i < len; i += 1) {
-    divs.push(<div key={i}/>)
+    divs.push(<div key={i} />);
   }
   ReactDOMServer.renderToString(
-    <IntlProvider locale="en">
-      {divs}
-    </IntlProvider>
+    <IntlProvider locale="en">{divs}</IntlProvider>
   );
 });
 
@@ -43,9 +42,7 @@ suite.add('100 x <FormattedNumber>', function() {
   }
 
   ReactDOMServer.renderToString(
-    <IntlProvider locale="en">
-      {formattedNumbers}
-    </IntlProvider>
+    <IntlProvider locale="en">{formattedNumbers}</IntlProvider>
   );
 });
 
@@ -57,9 +54,7 @@ suite.add('100 x <FormattedDate>', function() {
   }
 
   ReactDOMServer.renderToString(
-    <IntlProvider locale="en">
-      {formattedDates}
-    </IntlProvider>
+    <IntlProvider locale="en">{formattedDates}</IntlProvider>
   );
 });
 
@@ -82,8 +77,14 @@ suite.add('100 x <FormattedMessage> with placeholder', function() {
   let messages: Record<number, string> = {};
   let formattedMessages = [];
   for (let i = 0, len = 100; i < len; i += 1) {
-    const varName = `var${i}`
-    messages[i] = `{${varName}, plural, =0{{${varName}, number} message} =1{{${varName}, number} messages}}`;
+    const varName = `var${i}`;
+    messages[i] = `{${varName}, plural, 
+    zero {{${varName}, number} message} 
+    one {{${varName}, number} message} 
+    few {{${varName}, number} message} 
+    many {{${varName}, number} message} 
+    other {{${varName}, number} messages}
+  }`;
     formattedMessages.push(
       <FormattedMessage id={`${i}`} values={{[varName]: i}} key={i} />
     );
@@ -96,11 +97,48 @@ suite.add('100 x <FormattedMessage> with placeholder', function() {
   );
 });
 
+let messageAsts: Record<number, MessageFormatPattern> = {};
+for (let i = 0, len = 100; i < len; i += 1) {
+  const varName = `var${i}`;
+  messageAsts[i] = parser.parse(
+    `{${varName}, plural, 
+      zero {{${varName}, number} message} 
+      one {{${varName}, number} message} 
+      few {{${varName}, number} message} 
+      many {{${varName}, number} message} 
+      other {{${varName}, number} messages}
+    }`
+  );
+}
+
+suite.add('100 x <FormattedMessage> with placeholder in AST form', function() {
+  let formattedMessages = [];
+
+  for (let i = 0, len = 100; i < len; i += 1) {
+    const varName = `var${i}`;
+    formattedMessages.push(
+      <FormattedMessage id={`${i}`} values={{[varName]: i}} key={i} />
+    );
+  }
+
+  ReactDOMServer.renderToString(
+    <IntlProvider locale="en" messages={messageAsts}>
+      {formattedMessages}
+    </IntlProvider>
+  );
+});
+
 suite.add('100 x <FormattedMessage> with placeholder, cached', function() {
   let messages: Record<number, string> = {};
   let formattedMessages = [];
   for (let i = 0, len = 100; i < len; i += 1) {
-    messages[i] = `{var0, plural, =0{{var0, number} message} =1{{var0, number} messages}}`;
+    messages[i] = `{var0, plural, 
+      zero {{var0, number} message} 
+      one {{var0, number} message} 
+      few {{var0, number} message} 
+      many {{var0, number} message} 
+      other {{var0, number} messages}
+    }`;
     formattedMessages.push(
       <FormattedMessage id={`${i}`} values={{var0: i}} key={i} />
     );
@@ -112,6 +150,35 @@ suite.add('100 x <FormattedMessage> with placeholder, cached', function() {
     </IntlProvider>
   );
 });
+
+const cachedAst = parser.parse(
+  `{var0, plural, 
+    zero {{var0, number} message} 
+    one {{var0, number} message} 
+    few {{var0, number} message} 
+    many {{var0, number} message} 
+    other {{var0, number} messages}
+  }`
+);
+suite.add(
+  '100 x <FormattedMessage> with placeholder, cached in AST form',
+  function() {
+    let messages: Record<number, MessageFormatPattern> = {};
+    let formattedMessages = [];
+    for (let i = 0, len = 100; i < len; i += 1) {
+      messages[i] = cachedAst;
+      formattedMessages.push(
+        <FormattedMessage id={`${i}`} values={{var0: i}} key={i} />
+      );
+    }
+
+    ReactDOMServer.renderToString(
+      <IntlProvider locale="en" messages={messages}>
+        {formattedMessages}
+      </IntlProvider>
+    );
+  }
+);
 
 // suite.add('100 x <FormattedRelative>', function() {
 //   let formattedRelativeTimes = [];