--- a/test/perf/tsconfig.json
+++ b/test/perf/tsconfig.json
@@ -1,0 +1,4 @@
+{
+    "extends": "../../tsconfig.cjs.json",
+    "include": ["**/*.ts*"]
+}