depends on https://github.com/formatjs/formatjs/pull/107

- Passing in `formatters` from `context` allows us to share cache which can speed things up tremendously (uncached 3x speedup, cached normalized out)
- Fix the memory leak from calling `memoizeIntlConstructor` globally
- Fix benchmark script
- Allow `Provider` to render more than 1 children since React 16 supports that now

Before
```
<div> x 512,857 ops/sec ±1.28% (87 runs sampled)
100 x <div/> x 1,743 ops/sec ±2.65% (83 runs sampled)
100 x <FormattedNumber> x 670 ops/sec ±1.25% (88 runs sampled)
100 x <FormattedDate> x 399 ops/sec ±2.68% (84 runs sampled)
100 x <FormattedMessage> x 267 ops/sec ±5.50% (80 runs sampled)
100 x <FormattedMessage> with placeholder x 25.17 ops/sec ±3.14% (44 runs sampled)
100 x <FormattedMessage> with placeholder in AST form x 28.14 ops/sec ±2.30% (51 runs sampled)
100 x <FormattedMessage> with placeholder, cached x 329 ops/sec ±23.67% (78 runs sampled)
100 x <FormattedMessage> with placeholder, cached in AST form x 155 ops/sec ±1.25% (79 runs sampled)
```

After
```
<div> x 510,351 ops/sec ±1.87% (88 runs sampled)
100 x <div/> x 1,735 ops/sec ±2.65% (88 runs sampled)
100 x <FormattedNumber> x 647 ops/sec ±3.79% (84 runs sampled)
100 x <FormattedDate> x 433 ops/sec ±1.08% (87 runs sampled)
100 x <FormattedMessage> x 270 ops/sec ±4.88% (85 runs sampled)
100 x <FormattedMessage> with placeholder x 81.54 ops/sec ±3.64% (70 runs sampled)
100 x <FormattedMessage> with placeholder in AST form x 114 ops/sec ±1.65% (73 runs sampled)
100 x <FormattedMessage> with placeholder, cached x 372 ops/sec ±1.98% (86 runs sampled)
100 x <FormattedMessage> with placeholder, cached in AST form x 154 ops/sec ±1.86% (78 runs sampled)
```