@longlho 2019-07-02
Existing tests should continue to work. The rest is implementation details so I don't think docs are needed?